*** empty log message ***
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@96088 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 4e9a89b650342970201e9a1627d28879506ee8ea

+++ Online/Online/GaudiOnline/GaudiOnline/DimMessageSvc.h
@@ -1,4 +1,4 @@
-// $Header: /afs/cern.ch/project/cvs/reps/lhcb/Online/GaudiOnline/GaudiOnline/DimMessageSvc.h,v 1.1 2006-04-24 12:18:41 frankb Exp $
+// $Header: /afs/cern.ch/project/cvs/reps/lhcb/Online/GaudiOnline/GaudiOnline/DimMessageSvc.h,v 1.2 2006-04-24 14:45:05 frankb Exp $
 #ifndef GAUDIONLINE_DIMMESSAGESVC_H
 #define GAUDIONLINE_DIMMESSAGESVC_H
 
@@ -13,7 +13,7 @@ class DimService;
  *    LHCb namespace declaration
  */
 namespace LHCb  {
-  static const int SRCSIZE = 256;
+  static const size_t SRCSIZE = 256;
   struct DimMessage  {
     int typ, size;
     char src[SRCSIZE];
